backend/requirements/local.txt
-r base.txt
-r developement.txt

Werkzeug[watchdog]==2.0.3 # https://github.com/pallets/werkzeug
ipdb==0.13.9  # https://github.com/gotcha/ipdb
psycopg2==2.9.3  # https://github.com/psycopg/psycopg2
watchgod==0.8.2  # https://github.com/samuelcolvin/watchgod

# Testing
# ------------------------------------------------------------------------------
mypy==0.950  # https://github.com/python/mypy
django-stubs==1.9.0  # https://github.com/typeddjango/django-stubs
pytest==7.1.2  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.4  # https://github.com/Frozenball/pytest-sugar
djangorestframework-stubs==1.4.0  # https://github.com/typeddjango/djangorestframework-stubs

# Code quality
# ------------------------------------------------------------------------------
flake8==4.0.1  # https://github.com/PyCQA/flake8
# A plugin for Flake8 finding likely bugs and design problems in your program.
# https://github.com/PyCQA/flake8-bugbear
flake8-bugbear
# Extension to enforce better comma placement.
# https://github.com/PyCQA/flake8-commas
flake8-commas
# A flake8 plugin checking common style issues or inconsistencies with pytest-based tests.
# https://github.com/m-burst/flake8-pytest-style
flake8-pytest-style
# McCabe complexity checker
# https://github.com/PyCQA/mccabe
mccabe
coverage==6.3.2  # https://github.com/nedbat/coveragepy
black==22.3.0  # https://github.com/psf/black
pylint-django==2.5.3  # https://github.com/PyCQA/pylint-django
pylint-celery==0.3  # https://github.com/PyCQA/pylint-celery
pylint_pytest  # https://github__com.teameo.ca/carsongee/pytest-pylint
pylint_quotes  # https://github.com/edaniszewski/pylint-quotes
pre-commit==2.19.0  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.2.1  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==3.4.0  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.1.5  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==2.0.3  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.5.2  # https://github.com/pytest-dev/pytest-django
